configs/ropsten: fix params to support The Merge
Ropsten has moved to PoS and now needs a consensus client (such as
Prysm) to run alongside geth to function.

The merge didn't go without a hitch, the terminal difficulty needs to
be adjusted, see:

https://blog.ethereum.org/2022/06/03/ropsten-merge-ttd/

The http API flags are set according to these docs to allow Prysm to
work:

https://docs.prylabs.network/docs/install/install-with-script

The jwtsecret file is auto-generated by geth and will be used by Prysm
to authenticate.

Prysm connects on geth's authrpc port, which by default is 8551. We
set it to 18551 on Ropsten so that mainnet can run in parallel.